Fixed a couple errors and moved some utils files to the lib directory for better clarification
