+++ database/moves/hitboxes-data.sql
@@ -150,7 +150,7 @@ INSERT INTO public.hitboxes AS "captainFalcon"
     ('13.0/8.0%', '10-11/12-15', '35', '10'),
     ('10.0/9.0/8.0%', '7-10/11-12', '31', '7'),
     ('14.0%', '16-18', '44', '16'),
-    ('25.0/28.0 (22.0/25.0)%', '53-57 (62-66)', '103 (117)', '54'),
+    ('25.0/28.0 (22.0/25.0)%', '53-57 (62-66)', '103 (117)', '53'),
     ('10.0%', '5/10-25', '27/71', '5'),
     ('10.0%', '5/19-35', '44', '5'),
     ('5.0/13.0%', '14-30', null, '14'),
