+++ server/lib/handle-restart-id.ts
